refactor(desktop): move workspace file cache ownership
